+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [ ] Bower: Convert to new Azure based registry
 - [ ] Implement better auto-completion mechanism
+- [x] Setting to customize arguments for each package manager
 
 Features that have a checkmark are complete and available for
 download in the